Added withType to public API
Resolved  #53
